--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,5 +1,3 @@
-cargo-features = ["default-run"]
-
 [package]
 name = "kirby"
 version = "0.1.0"
@@ -12,7 +10,6 @@ default-run = "kirby"
 argparse = "0.2.2"
 aws_lambda = { git = "https://github.com/srijs/rust-aws-lambda" }
 flate2 = { version = "1.0", features = ["zlib"], default-features = false }
-hashbrown = { version = "0.1", features = ['serde', 'nightly'] }
 futures = "0.1.24"
 lazy_static = "1.1.0"
 enum-map = { version = "0.4.1", features = ["serde"] }