--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,15 +10,15 @@ default-run = "kirby"
 argparse = "0.2.2"
 aws_lambda = { git = "https://github.com/srijs/rust-aws-lambda" }
 flate2 = { version = "1.0", features = ["zlib"], default-features = false }
-futures = "0.1.24"
+futures = "0.3"
 lazy_static = "1.1.0"
 enum-map = { version = "0.4.1", features = ["serde"] }
 log = "0.4.5"
 nom = "4.0"
 rayon = "1.0.2"
 regex = "1"
-rusoto_core = { version="0.34.0", default_features=false, features=["rustls"] }
-rusoto_s3 = { version="0.34.0", default_features=false, features=["rustls"] }
+rusoto_core = { version="0.42", default_features=false, features=["rustls"] }
+rusoto_s3 = { version="0.42", default_features=false, features=["rustls"] }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"