--- a/src/user_agent.rs
+++ b/src/user_agent.rs
@@ -18,16 +18,16 @@ pub fn parse(a: &str) -> Option<UserAgent> {
   lazy_static! {
     // Here is the named regex. The regex created below does not include names, because that interface has borrowing issues 😬
     // \Abundler/(?<bundler>[0-9a-zA-Z.\-]+) rubygems/(?<rubygems>[0-9a-zA-Z.\-]+) ruby/(?<ruby>[0-9a-zA-Z.\-]+) \((?<platform>.*)\) command/(.*?)(?: jruby/(?<jruby>[0-9a-zA-Z.\-]+))?(?: truffleruby/(?<truffleruby>[0-9a-zA-Z.\-]+))?(?: options/(?<options>.*?))?(?: ci/(?<ci>.*?))? ([a-f0-9]{16})(?: Gemstash/(?<gemstash>[0-9a-zA-Z.\-]+))?\z
-    static ref br: Regex = Regex::new(r"\Abundler/([0-9a-zA-Z.\-]+) rubygems/([0-9a-zA-Z.\-]+) ruby/([0-9a-zA-Z.\-]+) \(([^)]*)\) command/(.*?)(?: jruby/([0-9a-zA-Z.\-]+))?(?: truffleruby/([0-9a-zA-Z.\-]+))?(?: options/(.*?))?(?: ci/(.*?))? [a-f0-9]{16}(?: Gemstash/([0-9a-zA-Z.\-]+))?\z").unwrap();
-    static ref rr: Regex = Regex::new(r"\A(?:Ruby, )?RubyGems/([0-9a-z.\-]+) (.*) Ruby/([0-9a-z.\-]+) \(.*?\)(?: jruby| truffleruby| rbx)?(?: Gemstash/([0-9a-z.\-]+))?\z").unwrap();
-    static ref gr: Regex = Regex::new(r"\ARuby, Gems ([0-9a-z.\-]+)\z").unwrap();
+    static ref BUNDLER_PATTERN: Regex = Regex::new(r"\Abundler/([0-9a-zA-Z.\-]+) rubygems/([0-9a-zA-Z.\-]+) ruby/([0-9a-zA-Z.\-]+) \(([^)]*)\) command/(.*?)(?: jruby/([0-9a-zA-Z.\-]+))?(?: truffleruby/([0-9a-zA-Z.\-]+))?(?: options/(.*?))?(?: ci/(.*?))? [a-f0-9]{16}(?: Gemstash/([0-9a-zA-Z.\-]+))?\z").unwrap();
+    static ref RUBY_PATTERN: Regex = Regex::new(r"\A(?:Ruby, )?RubyGems/([0-9a-z.\-]+) (.*) Ruby/([0-9a-z.\-]+) \(.*?\)(?: jruby| truffleruby| rbx)?(?: Gemstash/([0-9a-z.\-]+))?\z").unwrap();
+    static ref GEM_PATTERN: Regex = Regex::new(r"\ARuby, Gems ([0-9a-z.\-]+)\z").unwrap();
   }
 
-  let mut bl = br.capture_locations();
-  let mut rl = rr.capture_locations();
-  let mut gl = gr.capture_locations();
+  let mut bl = BUNDLER_PATTERN.capture_locations();
+  let mut rl = RUBY_PATTERN.capture_locations();
+  let mut gl = GEM_PATTERN.capture_locations();
 
-  if let Some(_) = br.captures_read(&mut bl, a) {
+  if let Some(_) = BUNDLER_PATTERN.captures_read(&mut bl, a) {
     return Some(UserAgent {
       bundler: match bl.get(1) {
         Some(loc) => Some(&a[loc.0..loc.1]),
@@ -70,7 +70,7 @@ pub fn parse(a: &str) -> Option<UserAgent> {
         _ => None,
       },
     });
-  } else if let Some(_) = rr.captures_read(&mut rl, a) {
+  } else if let Some(_) = RUBY_PATTERN.captures_read(&mut rl, a) {
     return Some(UserAgent {
       bundler: None,
       rubygems: match rl.get(1) {
@@ -95,7 +95,7 @@ pub fn parse(a: &str) -> Option<UserAgent> {
         _ => None,
       },
     });
-  } else if let Some(_) = gr.captures_read(&mut gl, a) {
+  } else if let Some(_) = GEM_PATTERN.captures_read(&mut gl, a) {
     return Some(UserAgent {
       bundler: None,
       rubygems: match gl.get(1) {