Also pulls in #19 and #20.