@indirect 2022-02-08
Before 9929759:

```
Benchmark 1: target/release/kirby logs/example.log.gz
  Time (mean ± σ):      1.810 s ±  0.012 s    [User: 1.793 s, System: 0.014 s]
  Range (min … max):    1.795 s …  1.830 s    10 runs
```

after [9929759](https://github.com/rubytogether/kirby/pull/21/commits/992975984336d4d914b1fe037ba64e8e287d502d)

```
Benchmark 1: target/release/kirby logs/example.log.gz
  Time (mean ± σ):      1.848 s ±  0.025 s    [User: 1.821 s, System: 0.018 s]
  Range (min … max):    1.819 s …  1.904 s    10 runs
```